+++ core-ng/src/main/java/core/framework/impl/template/expression/ExpressionBuilder.java
@@ -1,10 +1,15 @@
 package core.framework.impl.template.expression;
 
+import core.framework.api.util.Exceptions;
 import core.framework.api.util.Strings;
 import core.framework.impl.code.CodeBuilder;
 import core.framework.impl.code.CodeCompileException;
 import core.framework.impl.code.DynamicInstanceBuilder;
+import core.framework.impl.reflect.GenericTypes;
 import core.framework.impl.template.CallStack;
+
+import java.lang.reflect.Method;
+import java.lang.reflect.Type;
 
 /**
  * @author neo
@@ -22,13 +27,10 @@ public class ExpressionBuilder {
         token = new ExpressionParser().parse(expressionSource);
     }
 
-    public ExpressionWithSourceInfo build() {
+    public ExpressionHolder build() {
         Expression expression = buildExpression();
-        return new ExpressionWithSourceInfo(expression, expressionSource, location);
-    }
-
-    public Class<?> listValueClass() {
-        return new ExpressionTypeInspector().listValueClass(token, stack.rootClass, expressionSource);
+        Type returnType = returnType(token, stack.rootClass);
+        return new ExpressionHolder(expression, returnType, expressionSource, location);
     }
 
     private Expression buildExpression() {
@@ -53,5 +55,46 @@ public class ExpressionBuilder {
 
         builder.append("}");
         return builder.build();
+    }
+
+    private Type returnType(Token token, Class<?> modelClass) {
+        if (token instanceof MethodToken) {
+            if (((MethodToken) token).builtinMethod)
+                return Object.class;
+            String methodName = ((MethodToken) token).name;
+            Type returnType = methodReturnType(modelClass, methodName);
+            if (((MethodToken) token).next != null) {
+                return returnType(((MethodToken) token).next, GenericTypes.rawClass(returnType));
+            }
+            return returnType;
+        } else if (token instanceof FieldToken) {
+            Type fieldType = fieldType(modelClass, ((FieldToken) token).name);
+            if (((FieldToken) token).next != null) {
+                return returnType(((FieldToken) token).next, GenericTypes.rawClass(fieldType));
+            }
+            return fieldType;
+        } else {
+            return ((ValueToken) token).type;
+        }
+    }
+
+    private Type fieldType(Class<?> modelClass, String fieldName) {
+        Class<?> fieldClass = stack.paramClasses.get(fieldName);
+        if (fieldClass != null) return fieldClass;
+        try {
+            return modelClass.getField(fieldName).getGenericType();
+        } catch (NoSuchFieldException e) {
+            throw new Error(Strings.format("can not find field, class={}, field={}, expression={}, location={}",
+                modelClass, fieldName, expressionSource, location), e);
+        }
+    }
+
+    private Type methodReturnType(Class<?> modelClass, String methodName) {
+        Method[] methods = modelClass.getMethods();
+        for (Method method : methods) {
+            if (method.getName().equals(methodName)) return method.getGenericReturnType();
+        }
+        throw Exceptions.error("can not find method, class={}, method={}, expression={}, location={}",
+            modelClass, methodName, expressionSource, location);
     }
 }