+++ core-ng/src/main/java/core/framework/impl/template/expression/ValueToken.java
@@ -5,8 +5,10 @@ package core.framework.impl.template.expression;
  */
 public class ValueToken implements Token {
     final String value;
+    final Class<?> type;
 
-    public ValueToken(String value) {
+    public ValueToken(String value, Class<?> type) {
         this.value = value;
+        this.type = type;
     }
 }