+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.6.5 (10/08/2015)
 * web: support CDN config and #cdn() function, in sys.properties uses sys.cdn.host=
+* template: expression validate return type, for for/if
 
 ### 3.6.4 (10/07/2015)
 * validate: added @NotEmpty (for string), @ValueNotEmpty, @ValueNotNull (for collection)